Add additional rows, simplify orientation change detection
Note: On changing orientation, there's a temporary auto-layout conflict scenario as we show / hide scientific buttons.
That's because the UIStackView does not immediately generate the correct constraints when we toggle these button views' .isHidden property.
It does the right thing soon after though so the final layout is correct (and so is the animation).